\name{mixtNB-package}
\alias{mixtNB-package}
\alias{mixtNB}
\docType{package}
\title{DE Analysis of RNA-Seq Data by Mixtures of Negative Binomials}
\description{
A method for performing differential expression analysis of RNA-Seq data when replicates under two conditions are available is implemented. First, mixtures of Negative Binomial distributions are fitted on the data in order to estimate the dispersions, then the Wald test is performed. }
\details{
\tabular{ll}{
Package: \tab mixtNB\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-01\cr
License: \tab GPL-3\cr
}
}
\author{
Authros: Elisabetta Bonafede, Cinzia Viroli

Maintainer: Cinzia Viroli <cinzia.viroli@unibo.it>
}
\references{
E. Bonafede, F. Picard, S. Robin and C. Viroli (2015), Modelling overdispersion heterogeneity in differential expression analysis using mixtures, under revision. 
}
